Unify skill format validation rules
